app can convert from cartesian to cylindrical and spherical
